requirements.txt
async-timeout==5.0.1
bandit==1.8.3
bcrypt==4.1.2
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.0.2
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pbr==6.1.1
psycopg2-binary==2.9.9
Pygments==2.19.1
pymongo==4.6.2
python-dotenv==1.0.0
PyYAML==6.0.2
redis==5.0.1
rich==14.0.0
SQLAlchemy==2.0.40
stevedore==5.4.1
typing_extensions==4.13.2
Werkzeug==3.1.3
